[Refactor] Restructure example/test files: move golden to test, remove perf tests, English comments
- example_mla_decode_paged.py: remove golden_mla_decode_paged() (moved to test file), inline minimal PyTorch golden in _smoke_test for CI independence
- test_mla_decode_paged.py: add golden_mla_decode_paged() locally, remove run_bench/run_msprof/_prepare_bench_inputs, restrict argparse to l0/l1/l2/boundary/all
- Translate all Chinese comments to English
- Precision unchanged: L0 5/5 + L1 11/11 + Boundary 5/5, worst case l1_valrange_asym matched_ratio=0.99820 max_abs=7.81e-3
- ruff check + ruff format passed
- File sizes: example 544->485 lines, test 899->791 lines (-167 lines total)
